+++ lib/localFunction.js
@@ -1,13 +1,11 @@
 'use strict';
 
 var _ = require('lodash');
-var is = require('is');
 var request = require('request');
 
-var instance;
+var encodeFirestoreValue = require('./firestore/encodeFirestoreValue');
 
 var LocalFunction = function(trigger, urls, controller) {
-  instance = this;
   this.name = trigger.name;
   this.eventTrigger = trigger.eventTrigger;
   this.httpsTrigger = trigger.httpsTrigger;
@@ -24,11 +22,11 @@ var LocalFunction = function(trigger, urls, controller) {
 };
 
 LocalFunction.prototype._isDatabaseFunc = function(eventTrigger) {
-  return /firebase.database/.test(eventTrigger.eventType);
+  return _.includes(eventTrigger.eventType, 'firebase.database');
 };
 
 LocalFunction.prototype._isFirestoreFunc = function(eventTrigger) {
-  return /cloud.firestore/.test(eventTrigger.eventType);
+  return _.includes(eventTrigger.eventType, 'cloud.firestore');
 };
 
 LocalFunction.prototype._substituteParams = function(resource, params) {
@@ -48,140 +46,65 @@ LocalFunction.prototype._requestCallBack = function(err, response, body) {
   return console.log('\nRESPONSE RECEIVED FROM FUNCTION: ' + status + body);
 };
 
-LocalFunction.prototype._encodeFirestoreValue = function(data) {
-  var isPlainObject = function(input) {
-    return typeof input === 'object' &&
-      input !== null &&
-      _.isEqual(Object.getPrototypeOf(input), Object.prototype);
-  };
-
-  var encodeHelper = function(val) {
-    if (is.string(val)) {
-      return {
-        stringValue: val
-      };
-    }
-    if (is.boolean(val)) {
-      return {
-        booleanValue: val
-      };
-    }
-    if (is.integer(val)) {
-      return {
-        integerValue: val
-      };
-    }
-    // Integers are handled above, the remaining numbers are treated as doubles
-    if (is.number(val)) {
-      return {
-        doubleValue: val
-      };
-    }
-    if (is.date(val)) {
-      return {
-        timestampValue: val.toISOString()
-      };
-    }
-    if (is.array(val)) {
-      var encodedElements = [];
-      for (var i = 0; i < val.length; ++i) {
-        var enc = encodeHelper(val[i]);
-        if (enc) {
-          encodedElements.push(enc);
-        }
-      }
-      return {
-        arrayValue: {
-          values: encodedElements
-        }
-      };
-    }
-    if (is.nil(val)) {
-      return {
-        nullValue: 'NULL_VALUE'
-      };
-    }
-    if (is.instanceof(val, Buffer) || is.instanceof(val, Uint8Array)) {
-      return {
-        bytesValue: val
-      };
-    }
-    if (isPlainObject(val)) {
-      return {
-        mapValue: {
-          fields: instance._encodeFirestoreValue(val)
-        }
-      };
-    }
-    throw new Error(
-      'Cannot encode ' + val + 'to a Firestore Value.' +
-      ' The emulator does not yet support Firestore document reference values or geo points.'
-    );
-  };
-
-  return _.mapValues(data, encodeHelper);
-};
-
 LocalFunction.prototype._call = function(data, opts) {
   opts = opts || {};
   var operationType;
   var dataPayload;
+
+  var makeFirestoreValue = function(input) {
+    var currentTime = (new Date()).toISOString();
+    return {
+      fields: encodeFirestoreValue(input),
+      createTime: currentTime,
+      updateTime: currentTime
+    };
+  };
+
   if (this.httpsTrigger) {
     this.controller.call(this.name, data || {});
   } else if (this.eventTrigger) {
     if (this._isDatabaseFunc(this.eventTrigger)) {
       operationType = _.last(this.eventTrigger.eventType.split('.'));
-      if (operationType === 'create') {
+      switch (operationType) {
+      case 'create':
         dataPayload = {
           data: null,
           delta: data
         };
-      } else if (operationType === 'update' || operationType === 'write') {
-        dataPayload = {
-          data: data.before,
-          delta: data.after
-        };
-      } else if (operationType === 'delete') {
+        break;
+      case 'delete':
         dataPayload = {
           data: data,
           delta: null
+        };
+        break;
+      default: // 'update' or 'write'
+        dataPayload = {
+          data: data.before,
+          delta: data.after
         };
       }
       opts.resource = this._substituteParams(this.eventTrigger.resource, opts.params);
       this.controller.call(this.name, dataPayload, opts);
     } else if (this._isFirestoreFunc(this.eventTrigger)) {
       operationType = _.last(this.eventTrigger.eventType.split('.'));
-      var currentTime =  (new Date()).toISOString();
-      if (operationType === 'create') {
+      switch (operationType) {
+      case 'create':
         dataPayload = {
-          value: {
-            fields: this._encodeFirestoreValue(data),
-            createTime: currentTime,
-            updateTime: currentTime
-          },
+          value: makeFirestoreValue(data),
           oldValue: {}
         };
-      } else if (operationType === 'update' || operationType === 'write') {
-        dataPayload = {
-          value: {
-            fields: this._encodeFirestoreValue(data.after),
-            createTime: currentTime,
-            updateTime: currentTime
-          },
-          oldValue: {
-            fields: this._encodeFirestoreValue(data.before),
-            createTime: currentTime,
-            updateTime: currentTime
-          }
-        };
-      } else if (operationType === 'delete') {
+        break;
+      case 'delete':
         dataPayload = {
           value: {},
-          oldValue: {
-            fields: this._encodeFirestoreValue(data),
-            createTime: currentTime,
-            updateTime: currentTime
-          }
+          oldValue: makeFirestoreValue(data)
+        };
+        break;
+      default: // 'update' or 'write'
+        dataPayload = {
+          value: makeFirestoreValue(data.after),
+          oldValue: makeFirestoreValue(data.before)
         };
       }
       opts.resource = this._substituteParams(this.eventTrigger.resource, opts.params);
